chore: move extmock to tests
