+++ Cargo.toml
@@ -18,3 +18,6 @@ members = [
     "ewasm",
     "state"
 ]
+
+[features]
+test = ["evm/test"]
