+++ evm/src/lib.rs
@@ -1,8 +1,9 @@
 pub mod common;
 pub mod err;
 pub mod ext;
-pub mod extmock;
 pub mod interpreter;
 pub mod memory;
 pub mod opcodes;
 pub mod stack;
+#[cfg(feature = "test")]
+pub mod tests;
